Refactor av proptypes til separate filer.
