+++ src/felles-komponenter/medlemskap.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as Nav from '../utils/navFrontend';
+import * as MPT from '../proptypes/';
 
-import * as MPT from '../proptypes';
 import './medlemskap.css';
 import DatoOmrade from './datoOmrade/datoOmrade';
 import EnkeltDato from './datoOmrade/enkeltDato';
