+++ src/felles-komponenter/vilkarsveileder/vilkarsveileder.js
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router-dom';
 import PT from 'prop-types';
 
 import * as Ikon from '../../resources/images';
-import * as MPT from '../../proptypes';
+import * as MPT from '../../proptypes/';
 
 import StegLinje from './komponenter/stegLinje';
 import StegFane from './komponenter/stegFane';
