+++ src/proptypes/index.js
@@ -1,0 +1,39 @@
+import { Permisjonen, Permisjoner, PermisjonOgPermittering } from './permisjon';
+import { Inntekt, InntektLinje } from './inntekt';
+import { Arbeidsforhold, Arbeidsforholdene } from './arbeidsforhold';
+import { MedlemskapPeriode, Medlemskap } from './medlemskap';
+import { Arbeidsavtale } from './arbeidsavtale';
+import { Kodeverk } from './kodeverk';
+import { Person } from './person';
+import { BostedsAdresse } from './bosted';
+import { OrgnummerNavn, Organisasjon, Organisasjoner } from './organisasjon';
+import { SokListe, SokListeEnkeltlinje } from './sok';
+import { Saksbehandler } from './saksbehandler';
+import { Oppsummering } from './oppsummering';
+import { Periode } from './periode';
+import { Bekreftelser } from './bekreftelser';
+
+export {
+  Saksbehandler,
+  BostedsAdresse,
+  Person,
+  Periode,
+  Arbeidsavtale,
+  OrgnummerNavn,
+  PermisjonOgPermittering,
+  Arbeidsforhold,
+  Arbeidsforholdene,
+  Organisasjon,
+  Organisasjoner,
+  MedlemskapPeriode,
+  Medlemskap,
+  Bekreftelser,
+  Inntekt,
+  InntektLinje,
+  Oppsummering,
+  Permisjonen,
+  Permisjoner,
+  SokListe,
+  SokListeEnkeltlinje,
+  Kodeverk,
+};
